Data processing completed.

See merge request lab2526_summer/armv8_64!12

+++ src/bit_manipulation.h
@@ -8,7 +8,7 @@
  * Example: bitmask(3, 1) returns 0b1110.
  */
 static inline uint32_t bitmask(const uint8_t start, const uint8_t end) {
-    const int width = start - end + 1;
+    const uint8_t width = start - end + 1;
     if (width == 32) {
         return UINT32_MAX;
     }
@@ -19,7 +19,7 @@ static inline uint32_t bitmask(const uint8_t start, const uint8_t end) {
  * (Helper) Extracts bits start down to end from a value.
  * Example: extract_bits(3, 1, 0b10110) returns 0b011.
  */
-static inline uint64_t extract_bits(const uint8_t start, const uint8_t end, 
+static inline uint32_t extract_bits(const uint8_t start, const uint8_t end, 
                                     const uint64_t target) {
     return (bitmask(start, end) & target) >> end;
 }
@@ -31,8 +31,8 @@ static inline uint64_t extract_bits(const uint8_t start, const uint8_t end,
 static inline long long get_signed_value(const uint8_t start,
                                          const uint8_t end,
                                          const uint32_t target) {
-    const int width = start - end + 1;
-    const uint64_t unsigned_val = extract_bits(start, end, target);
+    const uint8_t width = start - end + 1;
+    const uint32_t unsigned_val = extract_bits(start, end, target);
     long long val = (long long)unsigned_val;
     if ((unsigned_val >> (width - 1)) & 1) {
         val -= 1LL << width;    // turns unsigned value into signed equivalent
